josh_talk/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
idna==3.11
requests==2.33.0
urllib3==2.6.3
torch
torchaudio
transformers
datasets
jiwer
librosa
soundfile
tqdm
numpy
pandas
scikit-learn
indic-nlp-library